Split admin commands to separate cog and endless to util
